Decompilation checkpoint: 1773120661

+++ src/decomp/sas_c/textdisp_handle_script_command.c
@@ -35,7 +35,7 @@ extern void MEMORY_DeallocateMemory(const char *file, LONG line, void *ptr, LONG
 extern LONG WDISP_SPrintf(char *dst, const char *fmt, const char *arg);
 extern LONG TEXTDISP_SelectGroupAndEntry(const char *cmd, char *primarySearch, LONG channelCode);
 extern UBYTE SCRIPT_GetBannerCharOrFallback(void);
-extern void TEXTDISP_BuildNowShowingStatusLine(LONG group, LONG match, LONG bannerChar);
+extern void TEXTDISP_BuildNowShowingStatusLine(unsigned short group, unsigned short match, unsigned short bannerChar);
 extern void SCRIPT_ResetBannerCharDefaults(void);
 extern void TEXTDISP_BuildEntryPairStatusLine(unsigned short group, unsigned short match, unsigned short bannerChar);
 extern void TEXTDISP_SetEntryTextFields(TEXTDISP_SelectionEntry *entry, const char *shortText, const char *longText);
